Tasks read operation via tasks module set
